feat: add centralized, cached access to the access token
